Refactor routing and authentication structure; remove old SignIn component and add new pages
