+++ frontend/src/Components/Dashboard.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function Dashboard() {
+  return (
+    <div>
+      <h1>Dashboard</h1>
+    </div>
+  )
+}
