+++ frontend/src/Components/Detailspages.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function Detailspages() {
+  return (
+    <div>
+        <h1>Details</h1>
+    </div>
+  )
+}
