+++ frontend/src/Components/Home.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function Home() {
+  return (
+    <div>
+      <h1>Home</h1>
+    </div>
+  )
+}
